requirements.txt
pandas
tqdm
uuid
python-dateutil